requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pyyaml>=6.0
httpx>=0.27
numpy>=1.26
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
